requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==4.0.1
python-dotenv==0.19.1
pyjwt==2.1.0
openai==0.27.0
requests==2.26.0
gunicorn==20.1.0
